+++ src/chads_davinci/__init__.py
@@ -6,7 +6,7 @@ chad.littlepage@gmail.com
 323.974.0444
 """
 
-__version__ = "0.2.5"
+__version__ = "0.2.6"
 __author__ = "Chad Littlepage"
 __email__ = "chad.littlepage@gmail.com"
 __phone__ = "323.974.0444"